requirements.txt
opencv-python>=4.10.0
numpy>=1.24.0
torch>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
faiss-cpu>=1.7.4
ultralytics>=8.2.0
